+++ adapters/aspnet/src/BugFab.AspNetCore/Data/Migrations/20260710232633_AddIdCounter.Designer.cs
@@ -3,6 +3,7 @@ using System;
 using BugFab.AspNetCore.Data;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 
 #nullable disable
@@ -10,9 +11,11 @@ using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 namespace BugFab.AspNetCore.Data.Migrations
 {
     [DbContext(typeof(BugFabDbContext))]
-    partial class BugFabDbContextModelSnapshot : ModelSnapshot
+    [Migration("20260710232633_AddIdCounter")]
+    partial class AddIdCounter
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder.HasAnnotation("ProductVersion", "8.0.10");